requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
tabulate==0.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
matplotlib==3.7.1
networkx==3.1
requests==2.31.0
python-dotenv==1.0.0